requirements.txt
Flask==2.2.3
plotly==5.8.0
python_dateutil==2.8.2
requests==2.27.1
scikit_learn==1.0.2
SQLAlchemy==1.4.36
torch==1.13.1
python-dotenv==1.0.0
gunicorn==20.1.0